Web site course
This is all the images for my website that we are building. The images attached are massive compared to the site and I called to ask about compressing the images and was told that we will be going over how to do that here in the next few steps and to just submit what I have so far.