Here is the webpage design
here is the webpage design we have been working on. It doesnt look just like in the video. Mostly because the most time i get to work on this is at night and there is no late night number for me to call when i need help so i had to problem solve and debug the written code.